+++ src/QNEthernet.cpp
@@ -125,7 +125,6 @@ void EthernetClass::setMACAddress(const uint8_t mac[6]) {
     return;
   }
 
-  bool dhcp = dhcpActive_;
   if (dhcpActive_) {
     dhcp_release_and_stop(netif_);  // Stop DHCP in all cases
     dhcpActive_ = false;
@@ -133,8 +132,7 @@ void EthernetClass::setMACAddress(const uint8_t mac[6]) {
 
   begin(netif_ip4_addr(netif_),
         netif_ip4_netmask(netif_),
-        netif_ip4_gw(netif_),
-        dhcp);
+        netif_ip4_gw(netif_));
 }
 
 void EthernetClass::loop() {
@@ -148,10 +146,6 @@ void EthernetClass::loop() {
 
 bool EthernetClass::begin() {
   return begin(INADDR_NONE, INADDR_NONE, INADDR_NONE);
-}
-
-bool EthernetClass::beginNoDHCP() {
-  return begin(nullptr, nullptr, nullptr, false);
 }
 
 bool EthernetClass::begin(const IPAddress &ip,
@@ -183,13 +177,12 @@ bool EthernetClass::begin(const IPAddress &ip,
   if (dns != INADDR_NONE) {
     setDNSServerIP(dns);
   }
-  return begin(&ipaddr, &netmask, &gw, true);
+  return begin(&ipaddr, &netmask, &gw);
 }
 
 bool EthernetClass::begin(const ip4_addr_t *ipaddr,
                           const ip4_addr_t *netmask,
-                          const ip4_addr_t *gw,
-                          bool canStartDHCP) {
+                          const ip4_addr_t *gw) {
   if (!enet_has_hardware()) {
     return false;
   }
@@ -222,12 +215,35 @@ bool EthernetClass::begin(const ip4_addr_t *ipaddr,
     // with any first subsequent DHCP requests
     // dhcp_inform(netif_);
     dhcpActive_ = false;
-  } else if (canStartDHCP) {
+    dhcpDesired_ = false;
+  } else if (dhcpEnabled_) {
     retval = (dhcp_start(netif_) == ERR_OK);
     dhcpActive_ = retval;
+    dhcpDesired_ = true;
   }
 
   attachLoopToYield();
+  return retval;
+}
+
+bool EthernetClass::setDHCPEnabled(bool flag) {
+  dhcpEnabled_ = flag;
+  if (netif_ == nullptr) {
+    return true;
+  }
+
+  bool retval = true;
+  if (flag) {  // DHCP enabled
+    if (dhcpDesired_ && !dhcpActive_) {
+      retval = (dhcp_start(netif_) == ERR_OK);
+      dhcpActive_ = retval;
+    }
+  } else {  // DHCP disabled
+    if (dhcpActive_) {
+      dhcp_release_and_stop(netif_);
+      dhcpActive_ = false;
+    }
+  }
   return retval;
 }
 
@@ -306,6 +322,7 @@ void EthernetClass::end() {
     dhcp_release_and_stop(netif_);
     dhcpActive_ = false;
   }
+  dhcpDesired_ = false;
 
   enet_deinit();
   netif_ = nullptr;